+++ src/vec/vec2.rs
@@ -176,8 +176,8 @@ macro_rules! vec2s {
             }
 
             #[inline]
-            pub fn map<F>(&self, f: F) -> Self
-                where F: Fn($t) -> $t
+            pub fn map<F>(&self, mut f: F) -> Self
+                where F: FnMut($t) -> $t
             {
                 $n::new(
                     f(self.x),
@@ -186,8 +186,8 @@ macro_rules! vec2s {
             }
 
             #[inline]
-            pub fn apply<F>(&mut self, f: F)
-                where F: Fn($t) -> $t
+            pub fn apply<F>(&mut self, mut f: F)
+                where F: FnMut($t) -> $t
             {
                 self.x = f(self.x);
                 self.y = f(self.y);